+++ packages/animated-pose/src/factory.ts
@@ -1,5 +1,6 @@
 import { Animated } from 'react-native';
 import poseFactory, { Poser, PoserConfig } from 'pose-core';
+import { warning } from 'hey-listen';
 import defaultTransitions from './inc/default-transitions';
 import { getUnit } from './inc/unit-conversion';
 import {
@@ -37,7 +38,12 @@ export default ({
   convertUnitToPoints,
   unitConverters
 }: AnimatedFactoryConfig): AnimatedPoserFactory => {
-  const pose = poseFactory<Value, Action, Action, AnimatedPoser>({
+  const pose: AnimatedPoserFactory = poseFactory<
+    Value,
+    Action,
+    Action,
+    AnimatedPoser
+  >({
     /**
      * Bind onChange callbacks
      */
@@ -77,11 +83,16 @@ export default ({
     createValue: (
       init,
       key,
-      { passiveParent, passiveProps, props }: CreateValueProps = {}
+      props,
+      { passiveParent, passiveProps, passiveParentKey }: CreateValueProps = {}
     ) => {
       if (passiveParent) {
         if (!nonLayoutValues.has(key)) {
           passiveParent.useNativeDriver = props.useNativeDriver = false;
+          warning(
+            false,
+            `useNativeDriver is invalidated on value "${passiveParentKey}", because interpolated value "${key}" can't be animated by the native driver.`
+          );
         }
         return { interpolation: passiveParent.raw.interpolate(passiveProps) };
       } else {
@@ -134,16 +145,30 @@ export default ({
      * If a transition has been defined as an object of props, convert this
      * into an Animated animation
      */
-    convertTransitionDefinition: ({ raw }, def, props) => {
+    convertTransitionDefinition: (
+      { raw, useNativeDriver },
+      def,
+      { toValue }
+    ) => {
       if (isAction(def)) return def;
 
+      const animationProps = {
+        ...def,
+        toValue,
+        useNativeDriver
+      };
+
       switch (def.type) {
-        case 'decay':
-          return Animated.decay(raw, def as Animated.DecayAnimationConfig);
         case 'spring':
-          return Animated.spring(raw, def as Animated.SpringAnimationConfig);
+          return Animated.spring(
+            raw,
+            animationProps as Animated.SpringAnimationConfig
+          );
         default:
-          return Animated.timing(raw, def as Animated.TimingAnimationConfig);
+          return Animated.timing(
+            raw,
+            animationProps as Animated.TimingAnimationConfig
+          );
       }
     },