+++ packages/animated-pose/src/types.ts
@@ -27,12 +27,12 @@ export type TransitionProps = {
 
 export type Transition = (
   { value, toValue }: TransitionProps
-) => Animated.CompositeAnimation;
+) => false | Animated.CompositeAnimation;
 
 export type CreateValueProps = {
   passiveParent?: Value;
   passiveProps?: Interpolation;
-  props?: { [key: string]: any };
+  passiveParentKey?: string;
 };
 
 export type AnimatedPoseConfig = PoserConfig<Value>;